Uploaded assigment 14 from application development class
In this assigment we were supposed to learn how to create and use new windows and pages in WPF framework. I took it step further and make decent logging app (decent for like 2 hours of programming, in reality this is bad app), that stores accounts in JSON file and hashes the password. App also customize itself with dark or white theme by what is your config in windows. (It uses registry value to determinate which theme to use.). More in README.md